DEV-004 make task links
DEV-004 This feature is testing the task link replacement functionality done by the github app. The Github app reads the pull request body and transforms ticket numbers found to app links. This works for all tasks links found in the body. For example: DEV-001 or DEV-005 will also be found and converted.